# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Artist": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"albums": {
						"items": {
							"$ref": "#/components/schemas/Album"
						},
						"type": "array"
					},
					"tracks": {
						"items": {
							"$ref": "#/components/schemas/Track"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"image",
					"albums",
					"tracks"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Album": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"views": {
						"type": "number",
						"format": "double"
					},
					"downloads": {
						"type": "number",
						"format": "double"
					},
					"release_date": {
						"type": "string",
						"format": "date-time"
					},
					"featured": {
						"type": "number",
						"format": "double"
					},
					"catalog": {
						"type": "string"
					},
					"top_album": {
						"type": "number",
						"format": "double"
					},
					"vinyl_album": {
						"type": "number",
						"format": "double"
					},
					"bandcamp_album": {
						"type": "number",
						"format": "double"
					},
					"artist": {
						"$ref": "#/components/schemas/Artist"
					},
					"category": {
						"$ref": "#/components/schemas/Category"
					},
					"publisher": {
						"$ref": "#/components/schemas/Publisher"
					},
					"language": {
						"$ref": "#/components/schemas/Language"
					},
					"tracks": {
						"items": {
							"$ref": "#/components/schemas/Track"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"title",
					"slug",
					"location",
					"views",
					"downloads",
					"release_date",
					"featured",
					"catalog",
					"top_album",
					"vinyl_album",
					"bandcamp_album",
					"artist",
					"category",
					"publisher",
					"language",
					"tracks",
					"created_at",
					"updated_at",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Track": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"duration": {
						"type": "number",
						"format": "double"
					},
					"bpm": {
						"type": "number",
						"format": "double"
					},
					"remixed": {
						"type": "string"
					},
					"filesize": {
						"type": "number",
						"format": "double"
					},
					"downloads": {
						"type": "number",
						"format": "double"
					},
					"search": {
						"type": "string"
					},
					"artist": {
						"$ref": "#/components/schemas/Artist"
					},
					"key": {
						"$ref": "#/components/schemas/Key"
					},
					"category": {
						"$ref": "#/components/schemas/Category"
					},
					"album": {
						"$ref": "#/components/schemas/Album"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"title",
					"slug",
					"location",
					"duration",
					"bpm",
					"remixed",
					"filesize",
					"downloads",
					"search",
					"artist",
					"key",
					"category",
					"album",
					"created_at",
					"updated_at",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Key": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"tracks": {
						"items": {
							"$ref": "#/components/schemas/Track"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"tracks"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Category": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"albums": {
						"items": {
							"$ref": "#/components/schemas/Album"
						},
						"type": "array"
					},
					"tracks": {
						"items": {
							"$ref": "#/components/schemas/Track"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"albums",
					"tracks"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Publisher": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"albums": {
						"items": {
							"$ref": "#/components/schemas/Album"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"albums"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Language": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"albums": {
						"items": {
							"$ref": "#/components/schemas/Album"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"albums"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IAlbumPayload": {
				"properties": {
					"title": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"artist_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"slug",
					"artist_id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-typescript",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/albums": {
			"get": {
				"operationId": "GetAlbums",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Album"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Album"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "skip",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "keyword",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "publisher",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "artist",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/albums/picked-albums": {
			"get": {
				"operationId": "GetPickedAlbums",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Album"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Album"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "type",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "publisher",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/albums/genre-albums/{slug}": {
			"get": {
				"operationId": "GetGenreAlbums",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Album"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Album"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "slug",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/albums/search": {
			"get": {
				"operationId": "GetSearchAlbums",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Album"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Album"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "keyword",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/albums/{slug}": {
			"get": {
				"operationId": "GetOneAlbum",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Album"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Album"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "slug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/albums/{id}": {
			"put": {
				"operationId": "UpdateAlbum",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Album"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Album"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAlbumPayload"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAlbum",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Album"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Album"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}